backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
aiohttp==3.9.1

# Presidio for PII detection and anonymization (pattern-based, no NLP)
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351

# MongoDB async driver
motor==3.3.2
pymongo==4.6.1

# Additional dependencies
python-multipart==0.0.6
